requirements.txt
pygame
python-decouple